Widen windows-sys bounds
The recent changes in windows-sys (https://github.com/microsoft/windows-rs/releases/tag/0.59.0) do not break etcetera (`cargo test` passes on both versions), but the strict bounds prevent downstream crates from upgrading, or as in our case, we get duplicate versions of windows-sys. Instead, we widen the bound to support both the old and the new version.